cloud-functions/pdf-converter/requirements.txt
google-cloud-storage==2.8.0
PyPDF2==3.0.1